server: Remove ServiceManager
Although it was easier to keep the sessions and other utility functions
stored inside the ServiceManager, storing them inside the Service makes more
sense. Because, the Service is responsible for handling them.
