+++ server/src/main.rs
@@ -3,7 +3,6 @@ mod error;
 mod item;
 mod prompt;
 mod service;
-mod service_manager;
 mod session;
 
 use clap::Parser;
